requirements.txt
Flask==2.0.1
google-cloud-storage==1.42.0
google-cloud-pubsub==2.9.0
fpdf2==2.5.5